% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mae.R
\name{calculate_mae}
\alias{calculate_mae}
\title{calculate_mae}
\usage{
calculate_mae(target, optimized_grid)
}
\arguments{
\item{target}{The target (commonness matrix) that was provided for the spectre optimization algorithm}

\item{optimized_grid}{The final grid (species X site) that was reported by the spectre algoithm}
}
\value{
mae real commonness error
}
\description{
The function calculates the real commonness error
}
\details{

}
\examples{
\dontrun{



}
}
