% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mc.R
\name{calculate_mc}
\alias{calculate_mc}
\title{calculate_mc}
\usage{
calculate_mc(target)
}
\arguments{
\item{target}{The target (commonness matrix) that was provided for the spectre optimization algorithm}
}
\value{
mc mean commonness
}
\description{
The function calculates the mean commonness of the target
}
\details{

}
\examples{
\dontrun{



}
}
