% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rce.R
\name{calculate_rce}
\alias{calculate_rce}
\title{calculate_rce}
\usage{
calculate_rce(target, optimized_grid)
}
\arguments{
\item{target}{The target (commonness matrix) that was provided for the spectre optimization algorithm}

\item{optimized_grid}{The final grid (species X site) that was reported by the spectre algoithm}
}
\value{
rce relative commonness error
}
\description{
The function calculates the relative commonness error
}
\details{

}
\examples{
\dontrun{



}
}
