% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualspecies_plot.R
\name{virtualspecies_plot}
\alias{virtualspecies_plot}
\title{virtualspecies_plot}
\usage{
virtualspecies_plot(species.stack, type = "richness")
}
\arguments{
\item{species.stack}{can either be a raster stack, or a named list of two raster stacks for side by side comparison}

\item{type}{either "richness" to plot the sum of the rasterstack, or "incidence" to plot all raster layers (species) as presences absence maps on a grid}
}
\value{
rasterstack
}
\description{
Plot virtual species raster stacks
}
\details{
Simple plotting functio for raster stacks.
If one raster stack is provided (for example generated by generate_data_virtualspecies()):
"richness" plots the sum of all layers (species) and "incidence" plots the prsence/absecne maps of all layers (species) on a grid.
If two raster stacks are provided (for example one generated by generate_data_virtualspecies() and one which was optimized through spectre):
"richness" plots the sum of both rasters ande the difference between both raster sums.
"incidence" does currently not work for two raster stacks.
}
\examples{
\dontrun{
ncol <- 5
nrow <- 5
gamma <- 100
corr_within <- 1 # strength of spatial autocorrelation within species
corr_among <- 1 # strength of spatial correlation among species (\% of gamma)
beta <- 0.75  # lower beta = wider species distributions

# Generate species and create list of alpha diversit:
spp <- generate_data_virtualspecies(ncol=ncol, nrow=nrow, corr_within = corr_within, corr_among = corr_among, gamma = gamma, beta = beta)

# plot species distributions:
virtualspecies_plot(spp, type="incidence")
# plot alpha diversity:
virtualspecies_plot(spp, type="richness")

}
}
